# docs/swagger.yaml
basePath: /
definitions:
  apierr.APIError:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  progress.AnswerRequest:
    properties:
      is_correct:
        type: boolean
    required:
    - is_correct
    type: object
  progress.ProgressResponse:
    properties:
      ease_factor:
        type: number
      interval_days:
        type: integer
      is_mastered:
        type: boolean
      is_saved:
        type: boolean
      last_reviewed_at:
        type: string
      next_review_at:
        type: string
      question_id:
        type: string
      repetitions:
        type: integer
    type: object
  questions.CodeChallengeResponse:
    properties:
      expected_output:
        type: string
      id:
        type: string
      language:
        type: string
      starter_code:
        type: string
      test_cases:
        type: string
    type: object
  questions.CreateOptionReq:
    properties:
      content:
        type: string
      is_correct:
        type: boolean
    required:
    - content
    type: object
  questions.CreateQuestionRequest:
    properties:
      content:
        type: string
      difficulty:
        enum:
        - beginner
        - intermediate
        - advanced
        type: string
      expected_output:
        type: string
      explanation:
        type: string
      language:
        type: string
      options:
        items:
          $ref: '#/definitions/questions.CreateOptionReq'
        type: array
      source:
        enum:
        - ai_generated
        - manual
        - imported
        type: string
      starter_code:
        type: string
      test_cases:
        type: string
      topic_ids:
        items:
          type: string
        minItems: 1
        type: array
      type:
        enum:
        - single_choice
        - multiple_choice
        - code_completion
        type: string
    required:
    - content
    - difficulty
    - topic_ids
    - type
    type: object
  questions.OptionResponse:
    properties:
      content:
        type: string
      id:
        type: string
      is_correct:
        type: boolean
    type: object
  questions.QuestionResponse:
    properties:
      code_challenge:
        $ref: '#/definitions/questions.CodeChallengeResponse'
      content:
        type: string
      created_at:
        type: string
      difficulty:
        type: string
      explanation:
        type: string
      id:
        type: string
      is_public:
        type: boolean
      options:
        items:
          $ref: '#/definitions/questions.OptionResponse'
        type: array
      source:
        type: string
      topics:
        items:
          $ref: '#/definitions/questions.TopicInfo'
        type: array
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  questions.TopicInfo:
    properties:
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
  questions.UpdateQuestionRequest:
    properties:
      content:
        type: string
      difficulty:
        enum:
        - beginner
        - intermediate
        - advanced
        type: string
      expected_output:
        type: string
      explanation:
        type: string
      language:
        type: string
      options:
        items:
          $ref: '#/definitions/questions.CreateOptionReq'
        type: array
      starter_code:
        type: string
      test_cases:
        type: string
      topic_ids:
        items:
          type: string
        type: array
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: felipe@example.com
    name: Felipe
  description: REST API con Go + Gin · Auth JWT · PostgreSQL · Swagger
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: dev-test-api
  version: "1.0"
paths:
...
  /api/v1/progress/{question_id}/answer:
    post:
      consumes:
      - application/json
      description: Registra una respuesta correcta/incorrecta y actualiza el progreso
        SM-2
      parameters:
      - description: Question ID
        in: path
        name: question_id
        required: true
        type: string
      - description: Respuesta
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/progress.AnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/progress.ProgressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierr.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Responder pregunta
      tags:
      - progress
  /api/v1/progress/{question_id}/toggle-save:
    post:
      description: Alterna el estado is_saved de una pregunta para el usuario
      parameters:
      - description: Question ID
        in: path
        name: question_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/progress.ProgressResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Marcar/desmarcar pregunta
      tags:
      - progress
  /api/v1/progress/saved:
    get:
      description: Lista todas las preguntas que el usuario ha guardado para repasar
      parameters:
      - description: 'Número de página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Elementos por página (default: 20, max: 100)'
        in: query
        name: per_page
        type: integer
      - description: 'Campo de ordenación: updated_at, repetitions, ease_factor'
        in: query
        name: sort_by
        type: string
      - description: 'Dirección: asc o desc (default: desc)'
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Preguntas guardadas (con paginación)
          schema:
            $ref: '#/definitions/response.Meta'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Preguntas guardadas (con paginación)
      tags:
      - progress
  /api/v1/progress/upcoming:
    get:
      description: Lista las preguntas marcadas para repasar cuyo próximo repaso es
        hoy
      parameters:
      - description: 'Número de página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Elementos por página (default: 20, max: 100)'
        in: query
        name: per_page
        type: integer
      - description: 'Campo de ordenación: next_review_at, repetitions, ease_factor'
        in: query
        name: sort_by
        type: string
      - description: 'Dirección: asc o desc (default: asc)'
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Preguntas pendientes (con paginación)
          schema:
            $ref: '#/definitions/response.Meta'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Preguntas pendientes (con paginación)
      tags:
      - progress
  /api/v1/questions:
    get:
      description: Lista todas las preguntas disponibles, con filtros opcionales
      parameters:
      - description: 'Número de página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Elementos por página (default: 20, max: 100)'
        in: query
        name: per_page
        type: integer
      - description: Filtrar por tipo (single_choice, multiple_choice, code_completion)
        in: query
        name: type
        type: string
      - description: Filtrar por dificultad (beginner, intermediate, advanced)
        in: query
        name: difficulty
        type: string
      - description: 'Campo de ordenación: type, difficulty, created_at, updated_at'
        in: query
        name: sort_by
        type: string
      - description: 'Dirección: asc o desc (default: desc)'
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de preguntas (con paginación)
          schema:
            $ref: '#/definitions/response.Meta'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Listar preguntas (con paginación)
      tags:
      - questions
    post:
      consumes:
      - application/json
      description: Crea una nueva pregunta con opciones o código
      parameters:
      - description: Datos de la pregunta
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/questions.CreateQuestionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/questions.QuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierr.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Crear pregunta (Admin)
      tags:
      - questions
  /api/v1/questions/{id}:
    delete:
      description: Elimina una pregunta (soft delete)
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierr.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Eliminar pregunta (Admin)
      tags:
      - questions
    get:
      description: Obtiene una pregunta por ID con todos sus detalles (opciones, código,
        temas)
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/questions.QuestionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Obtener pregunta
      tags:
      - questions
    put:
      consumes:
      - application/json
      description: Actualiza una pregunta existente
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/questions.UpdateQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/questions.QuestionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierr.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierr.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Actualizar pregunta (Admin)
      tags:
      - questions
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Prefix with "Bearer ". Example: "Bearer eyJhbG..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"